% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-leslies.R
\name{adjust_lambda}
\alias{adjust_lambda}
\title{Adjust Leslie matrix to match a given overall growth rate (lambda)}
\usage{
adjust_lambda(
  L,
  lambda,
  adjust_part = "f",
  max_val = NA,
  direct_l = FALSE,
  precision = 1e-04,
  optim_tol = .Machine$double.eps^0.5
)
}
\arguments{
\item{L}{Square numeric Leslie matrix that will be adjusted.}

\item{lambda}{Single numeric indicating lambda that output matrix will
be adjusted to.}

\item{adjust_part}{Single character indicating which part of the Leslie
matrix to adjust. Options are survivals (\code{"s"}) or fecundities (\code{"f"}).
Defaults to \code{"f"}.}

\item{max_val}{Single numeric indicating the maximum value of the modifier
to use to achieve the desired lambda.
Defaults to \code{NA}, which results in a very coarse search (by powers of
10) for a maximum.}

\item{direct_l}{Single logical for whether to use the \code{DIRECT_L} global
optimizer from \code{nloptr}. It's usually more accurate, but slower.
Defaults to \code{FALSE}.}

\item{precision}{Single numeric indicating how close the output lambda must
be to the desired value.
Defaults to \code{1e-4}.}

\item{optim_tol}{Single numeric indicating the tolerance to use for the
optimization process.
If \code{direct_l = FALSE}, this is the \code{tol} argument for \code{stats::optimize}.
If \code{direct_l = TRUE}, this is used for \code{xtol_rel} in NLOPT options
(see \code{?nl.opts}).
Defaults to \code{.Machine$double.eps^0.5}.}
}
\value{
An adjusted Leslie matrix.
}
\description{
Adjust Leslie matrix to match a given overall growth rate (lambda)
}
