% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphidsync-package.R
\docType{package}
\name{aphidsync-package}
\alias{aphidsync}
\alias{aphidsync-package}
\title{aphidsync: Simulate and Estimate Aphid Synchrony}
\description{
Simulate aphid population dynamics and estimate synchrony.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucasnell/aphidsync}
  \item Report bugs at \url{https://github.com/lucasnell/aphidsync/issues}
}

}
\author{
\strong{Maintainer}: Lucas A. Nell \email{lucnell@gmail.com} (\href{https://orcid.org/0000-0003-3209-0517}{ORCID}) [copyright holder]

}
\keyword{internal}
