% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_lambda}
\alias{calc_lambda}
\title{Calculate lambda (exponential growth rate) from Leslie matrix.}
\usage{
calc_lambda(L)
}
\arguments{
\item{L}{Leslie matrix.}
}
\description{
Calculate lambda (exponential growth rate) from Leslie matrix.
}
