% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known-leslie.R
\name{fit_known_leslie}
\alias{fit_known_leslie}
\title{Fit starting abundance for known Leslie matrix (optionally with uncertainty)}
\usage{
fit_known_leslie(
  sim_df,
  L,
  N0,
  K = NA_real_,
  adjust_L = 0L,
  max_shape = 800,
  compare_N = FALSE,
  fit_max_t = 30L,
  est_K_start = 50L,
  est_K_min_n = 50L,
  max_fecund_x = 100,
  optim_args = list()
)
}
\arguments{
\item{sim_df}{A dataframe containing the time and abundance columns.
Requires columns \code{"time"} and \code{"N"}.}

\item{L}{Leslie matrix used for fitting.}

\item{N0}{Starting abundance.}

\item{K}{Density dependence if known, or \code{NA} if unknown. Defaults to \code{NA}.}

\item{adjust_L}{Single integer for whether and how to fit parameters
that adjust the Leslie matrix.
A value of \code{0L} results in no Leslie adjustment,
\code{1L} adjusts the Leslie using 1 parameter (i.e., \code{L = L * leslie_x}),
and \code{2L} adjusts the Leslie using separate parameters for
the fecundities and survivals.
Fecundities (\code{f}) are adjusted as \code{f = f * fecund_x}.
Survivals (\code{s}) are adjusted via \code{s = inv_logit(logit(s) + surv_x)}.
Defaults to \code{0L}.}

\item{max_shape}{Single numeric for the maximum value for the shape
parameter. Depending on the number of stages, greater increases to
the shape parameter have little effect. Defaults to \code{800}.}

\item{compare_N}{Single logical for whether to use \code{N} or per-capita growth
when comparing between input data and object function simulations.}

\item{fit_max_t}{Single integer for the number of time steps to use
for estimating the agreement between observed vs fitted values.
Too high of a value here results in density dependence coming into
play more strongly. Defaults to \code{30L}.}

\item{est_K_start}{Single integer for the time at which to start estimating
density dependence. Defaults to \code{50L}.}

\item{est_K_min_n}{Single integer for the number of time steps to use
for estimating density dependence. Defaults to \code{50L}.}

\item{max_fecund_x}{Single numeric for the maximum value of \code{fecund_x} to
allow. Parameter \code{fecund_x} is the value that gets multiplied by
fecundities when adjusting the Leslie matrix.
Defaults to \code{100}.}

\item{optim_args}{List containing arguments to pass to \code{winnowing_optim}.
Defaults to \code{list()}.}
}
\value{
A named vector indicating the parameter values from the fit.
}
\description{
Fit starting abundance for known Leslie matrix (optionally with uncertainty)
}
