% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{med_age}
\alias{med_age}
\title{Median age}
\usage{
med_age(shape, offset, n_stages = 29L)
}
\arguments{
\item{shape}{Vector of shape parameters for the symmetrical beta
distribution of abundances.}

\item{offset}{Vector of offset value(s) for the symmetrical beta
distribution of abundances.
Must be the same length as \code{shape}.}
}
\description{
Median age
}
