% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-leslies.R
\name{random_fecunds}
\alias{random_fecunds}
\title{Perturb Leslie matrix fecundities stochastically}
\usage{
random_fecunds(
  L,
  sigma_f,
  norm_rnd = FALSE,
  norm_lambda = FALSE,
  al_args = list()
)
}
\arguments{
\item{L}{A numeric, square Leslie matrix.}

\item{sigma_f}{Standard deviation for the normal distribution used to
generate deviates that are exponentiated (i.e., \code{exp(x)} for random
deviate(s) \code{x}), then multiplied by fecundities.}

\item{norm_rnd}{Single logical for whether random variables are
normalized so that they always have a mean of one. This results in
fecundities for the output Leslie matrix that always sum to the
same number as those for the input matrix.
This does NOT mean that the growth rate of the Leslie matrix will
be the same.}

\item{norm_lambda}{Single logical for whether the output survivals
are normalized so that the output Leslie matrix has the same growth rate
as the input.
Defaults to \code{FALSE}.}

\item{al_args}{Named list containing arguments to pass to \code{\link[=adjust_lambda]{adjust_lambda()}}
that is used if \code{norm_lambda = TRUE}. The \code{adjust_lambda} arguments
\code{L}, \code{lambda}, and \code{adjust_part} are ignored if provided here since
they are defined inside this function.
Defaults to \code{list()}.}
}
\value{
A Leslie matrix with adjusted fecundities.
}
\description{
Perturb Leslie matrix fecundities stochastically
}
