% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-leslies.R
\name{random_survs}
\alias{random_survs}
\title{Perturb Leslie matrix survivals stochastically}
\usage{
random_survs(
  L,
  sigma_s,
  norm_rnd = FALSE,
  norm_lambda = FALSE,
  al_args = list()
)
}
\arguments{
\item{L}{A numeric, square Leslie matrix.}

\item{sigma_s}{Standard deviation for the normal distribution used to
generate deviates that are added to the logit-transformed survivals,
before being back-transformed.}

\item{norm_rnd}{Single logical for whether random variables are normalized
so that they always have a mean of zero. This results in
survivals for the output Leslie matrix that always sum to the
same number in logit-transformed space as those for the input matrix.
This does NOT mean that the growth rate of the Leslie matrix will
be the same.}

\item{norm_lambda}{Single logical for whether the output survivals
are normalized so that the output Leslie matrix has the same growth rate
as the input.
Defaults to \code{FALSE}.}

\item{al_args}{Named list containing arguments to pass to \code{\link[=adjust_lambda]{adjust_lambda()}}
that is used if \code{norm_lambda = TRUE}. The \code{adjust_lambda} arguments
\code{L}, \code{lambda}, and \code{adjust_part} are ignored if provided here since
they are defined inside this function.
Defaults to \code{list()}.}
}
\value{
A Leslie matrix with adjusted survivals.
}
\description{
Perturb Leslie matrix survivals stochastically
}
