% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{width99}
\alias{width99}
\title{Width of 99th quartile}
\usage{
width99(shape)
}
\arguments{
\item{shape}{Vector of shape parameters for the symmetrical beta
distribution of abundances.}
}
\description{
Width of 99th quartile
}
