% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizing.R
\name{winnowing_optim}
\alias{winnowing_optim}
\title{Winnowing optimization}
\usage{
winnowing_optim(
  fn,
  lower_bounds,
  upper_bounds,
  fn_args = list(),
  box_control = list(maxit = 100, reltol = 1e-04),
  fine_control = list(maxit = 500, reltol = 1e-06),
  polished_control = list(maxit = 1000, reltol = 1e-08),
  box_optim = optim,
  fine_optim = optim,
  polished_optim = optim,
  n_bevals = 100L,
  n_boxes = 1000L,
  n_fine = 100L,
  n_polished = 20L,
  extra_optims = 10L
)
}
\arguments{
\item{fn}{Objective function to minimize.}

\item{lower_bounds}{Lower bounds of boxes for each parameter.}

\item{upper_bounds}{Upper bounds of boxes for each parameter.}

\item{fn_args}{List containing other arguments to use for \code{fn}.
Defaults to \code{list()}.}

\item{box_control}{List containing arguments to use for the \code{control}
argument for the optimization that occurs for each box.
Defaults to \code{list(maxit = 100, reltol = 1e-4)}
(these names are adjusted if using an optimizer other than \code{optim}).}

\item{fine_control}{List containing arguments to use for the \code{control}
argument for the fine optimizations.
Defaults to \code{list(maxit = 500, reltol = 1e-6)}
(these names are adjusted if using an optimizer other than \code{optim}).}

\item{polished_control}{List containing arguments to use for the \code{control}
argument for the polished optimizations.
Defaults to \code{list(maxit = 1000, reltol = 1e-8)}
(these names are adjusted if using an optimizer other than \code{optim}).}

\item{box_optim}{Function to use for the optimization for each box.
Must be \code{stats::optim} or a function from packages \code{minqa} or \code{nloptr}.
Defaults to \code{stats::optim}.}

\item{fine_optim}{Function to use for the fine optimizations.
Must be \code{stats::optim} or a function from packages \code{minqa} or \code{nloptr}.
Defaults to \code{stats::optim}.}

\item{polished_optim}{Function to use for the polished optimizations.
Must be \code{stats::optim} or a function from packages \code{minqa} or \code{nloptr}.
Defaults to \code{stats::optim}.}

\item{n_bevals}{Number of evaluations of \code{fn} per box. Defaults to \code{100L}.}

\item{n_boxes}{Number of boxes. Defaults to \code{1000L}.}

\item{n_fine}{Number of fine optimizations.
Must be \verb{> n_boxes} and \verb{< n_polished}.
Defaults to \code{100L}.}

\item{n_polished}{Number of polished optimizations.
Must be \verb{> n_fine}.
Defaults to \code{20L}.}

\item{extra_optims}{Number of potential extra rounds of optimizations to run
if the best polished optimization still hasn't converged.
Defaults to \code{10L}.}
}
\value{
A single object of the class returned by \code{polished_optim}.
}
\description{
Winnowing optimization
}
