% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known-leslie.R
\name{fit_known_leslie}
\alias{fit_known_leslie}
\title{Fit starting abundance for known Leslie matrix (optionally with uncertainty)}
\usage{
fit_known_leslie(
  sim_df,
  L,
  N0,
  K = NA_real_,
  adjust_L = 0L,
  max_shape = 800,
  compare_N = FALSE,
  fit_max_t = 30L,
  est_K_start = 50L,
  est_K_min_n = 50L,
  max_fecund_x = 10,
  max_surv_x = 10,
  trans_base = 1.36,
  optim_args = list(),
  return_optims = FALSE
)
}
\arguments{
\item{sim_df}{A dataframe containing the time and abundance columns.
Requires columns \code{"time"} and \code{"N"}.}

\item{L}{Leslie matrix used for fitting.}

\item{N0}{Starting abundance.}

\item{K}{Density dependence if known, or \code{NA} if unknown. Defaults to \code{NA}.}

\item{adjust_L}{Single integer for whether and how to fit parameters
that adjust the Leslie matrix.
A value of \code{0L} results in no Leslie adjustment,
\code{1L} adjusts the Leslie using 1 parameter (i.e., \code{L[,1] = L[,1] * fecund_x}),
and \code{2L} adjusts the Leslie using separate parameters for
the fecundities and survivals.
Fecundities (\code{f}) are adjusted as \code{f = f * fecund_x}.
Survivals (\code{s}) are adjusted via \code{s = inv_logit(logit(s) + surv_x)}.
Defaults to \code{0L}.}

\item{max_shape}{Single numeric for the maximum value for the shape
parameter. Depending on the number of stages, greater increases to
the shape parameter have little effect. Defaults to \code{800}.}

\item{compare_N}{Single logical for whether to use \code{N} or per-capita growth
when comparing between input data and object function simulations.}

\item{fit_max_t}{Single integer for the number of time steps to use
for estimating the agreement between observed vs fitted values.
Too high of a value here results in density dependence coming into
play more strongly. Defaults to \code{30L}.}

\item{est_K_start}{Single integer for the time at which to start estimating
density dependence. Defaults to \code{50L}.}

\item{est_K_min_n}{Single integer for the number of time steps to use
for estimating density dependence. Defaults to \code{50L}.}

\item{max_fecund_x}{Single numeric for the maximum value of \code{fecund_x} to
search for within \code{winnowing_optim}.
Parameter \code{fecund_x} is the value that gets multiplied by
fecundities when adjusting the Leslie matrix using
2 parameters (i.e., \code{adjust_L = 2L}). Ignored if \code{adjust_L != 2L}.
Defaults to \code{10}.}

\item{max_surv_x}{Single numeric for the maximum absolute value of \code{surv_x}
to search for within \code{winnowing_optim}.
Parameter \code{surv_x} is the value that gets added to
logit-transformed survivals when adjusting the Leslie matrix using
2 parameters (i.e., \code{adjust_L = 2L}). Ignored if \code{adjust_L != 2L}.
Defaults to \code{10}.}

\item{trans_base}{Single numeric indicating the base to use for
transforming the shape and (optionally) \code{fecund_x} parameters when
fitting.
Each parameter \code{z} will be transformed as \code{trans_base^z} inside the
objective function, and back-transformed as \code{log(z, trans_base)}
in this one.
Using \code{exp(1)} here seems to cause the optimizer to have a hard time.
The default for this is \code{1.36}, which is about \code{exp(1) / 2}.}

\item{optim_args}{List containing arguments to pass to \code{winnowing_optim}.
Defaults to \code{list()}.}

\item{return_optims}{Single logical for whether to return all polished
optimization objects from \code{winnowing_optim}.
The alternative is to just return the fitted parameter values with
a vector indicating the greatest divergence between any of the polished
fits for each parameter.
Defaults to \code{FALSE}.}
}
\value{
A named vector indicating the parameter values from the fit.
}
\description{
Fit starting abundance for known Leslie matrix (optionally with uncertainty)
}
